+++ amfiteatr_core/src/env/traits/tracing_env.rs
@@ -1,9 +1,9 @@
-use crate::env::{EnvironmentStateSequential,  GameTrajectory};
+use crate::env::{SequentialGameState, GameTrajectory};
 use crate::domain::DomainParameters;
 
 
 /// Environment that provide tracing game.
-pub trait TracingEnvironment<DP: DomainParameters, S: EnvironmentStateSequential<DP>>{
+pub trait TracingEnvironment<DP: DomainParameters, S: SequentialGameState<DP>>{
 
     
     fn trajectory(&self) -> &GameTrajectory<DP, S>;